requirements.txt
pandas
mysql-connector-python
sqlalchemy
python-dotenv
google-cloud-bigquery
pyarrow